Move get_resource_pk_filter to its own
